requirements.txt
fastapi
uvicorn
openai-whisper
moviepy
ffmpeg
python-multipart
aiofiles
